Refactor: directory of services changed
