+++ src/components/servicesContainer/services/serviceCard/serviceDetails/ServiceDetails.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 
-export default class DetalhesServ extends Component {
+export default class ServiceDetails extends Component {
   render() {
     return (
       <div>
